fixed small error in accomodations.py

+++ server/models/accommodations.py
@@ -4,7 +4,7 @@ from .base_case import base_case as Base
 class Accommodation(Base):
     __tablename__ = "accommodations"
 
-    id = Column(Integer(15), primary_key=True, unique=True, nullable=False)
+    id = Column(Integer, primary_key=True, unique=True, nullable=False)
     name = Column(String(20), nullable=False)
     address = Column(String(40), nullable=False)
     city = Column(String(15), nullable=False)
